msm: kgsl: Convert all internal GPU addresses to 64 bit
Use 64 bit types for all internal references to GPU memory addresses
or sizes.

+++ drivers/gpu/msm/adreno_compat.c
@@ -64,9 +64,14 @@ int adreno_getproperty_compat(struct kgsl_device *device,
 				/*
 				 * NOTE: with mmu enabled, gpuaddr doesn't mean
 				 * anything to mmap().
+				 * NOTE: shadowprop.gpuaddr is uint32
+				 * (because legacy) and the memstore gpuaddr is
+				 * 64 bit. Cast the memstore gpuaddr to uint32.
 				 */
-				shadowprop.gpuaddr = device->memstore.gpuaddr;
-				shadowprop.size = device->memstore.size;
+				shadowprop.gpuaddr =
+					(unsigned int) device->memstore.gpuaddr;
+				shadowprop.size =
+					(unsigned int) device->memstore.size;
 				/*
 				 * GSL needs this to be set, even if it
 				 * appears to be meaningless
